+++ src/schemathesis/specs/openapi/schemas.py
@@ -2,8 +2,7 @@ from __future__ import annotations
 
 import itertools
 import string
-from collections import defaultdict
-from contextlib import ExitStack, contextmanager, suppress
+from contextlib import contextmanager, suppress
 from dataclasses import dataclass
 from difflib import get_close_matches
 from json import JSONDecodeError
@@ -43,7 +42,6 @@ from schemathesis.openapi.checks import JsonSchemaError, MissingContentType
 from schemathesis.specs.openapi import adapter
 from schemathesis.specs.openapi.adapter import OpenApiResponses, prepare_parameters
 from schemathesis.specs.openapi.adapter.protocol import SpecificationAdapter
-from schemathesis.specs.openapi.stateful import links
 from schemathesis.specs.openapi.types import v3
 
 from ...generation import GenerationMode
@@ -61,7 +59,7 @@ from .parameters import (
     OpenAPI30Parameter,
     OpenAPIParameter,
 )
-from .references import ConvertingResolver, ReferenceResolver
+from .references import ReferenceResolver
 from .security import BaseSecurityProcessor, OpenAPISecurityProcessor, SwaggerSecurityProcessor
 from .stateful import create_state_machine
 
@@ -104,10 +102,7 @@ def get_template_fields(template: str) -> set[str]:
 
 @dataclass(eq=False, repr=False)
 class BaseOpenAPISchema(BaseSchema):
-    links_field: ClassVar[str] = ""
-    header_required_field: ClassVar[str] = ""
     security: ClassVar[BaseSecurityProcessor] = None  # type: ignore
-    component_locations: ClassVar[tuple[tuple[str, ...], ...]] = ()
     _path_parameter_template: ClassVar[dict[str, Any]] = None  # type: ignore
     adapter: SpecificationAdapter = None  # type: ignore
 
@@ -155,7 +150,7 @@ class BaseOpenAPISchema(BaseSchema):
                 method="",
                 path="",
                 label="",
-                definition=OperationDefinition(raw=None, scope=""),
+                definition=OperationDefinition(raw=None),
                 schema=None,  # type: ignore
                 responses=None,  # type: ignore
             )
@@ -184,7 +179,7 @@ class BaseOpenAPISchema(BaseSchema):
         resolve = self.resolver.resolve
         resolve_path_item = self._resolve_path_item
         should_skip = self._should_skip
-        links_field = self.links_field
+        links_keyword = self.adapter.links_keyword
 
         # For operationId lookup
         selected_operations_by_id: set[str] = set()
@@ -211,7 +206,7 @@ class BaseOpenAPISchema(BaseSchema):
                         for response in definition.get("responses", {}).values():
                             if "$ref" in response:
                                 _, response = resolve(response["$ref"])
-                            defined_links = response.get(links_field)
+                            defined_links = response.get(links_keyword)
                             if defined_links is not None:
                                 statistic.links.total += len(defined_links)
                                 if is_selected:
@@ -401,7 +396,7 @@ class BaseOpenAPISchema(BaseSchema):
         operation: APIOperation[OpenAPIParameter, OpenApiResponses] = APIOperation(
             path=path,
             method=method,
-            definition=OperationDefinition(definition, scope),
+            definition=OperationDefinition(definition),
             base_url=base_url,
             app=self.app,
             schema=self,
@@ -507,28 +502,8 @@ class BaseOpenAPISchema(BaseSchema):
     def _get_parameter_serializer(self, definitions: list[dict[str, Any]]) -> Callable | None:
         raise NotImplementedError
 
-    def get_headers(
-        self, operation: APIOperation, response: Response
-    ) -> tuple[list[str], dict[str, dict[str, Any]] | None] | None:
-        definition = operation.responses.find_by_status_code(response.status_code)
-        if definition is None:
-            return None
-        # TODO: It should be proper scopes / resolve it eagerly
-        return [], definition.definition.get("headers")
-
     def as_state_machine(self) -> type[APIStateMachine]:
         return create_state_machine(self)
-
-    def get_links(self, operation: APIOperation) -> dict[str, dict[str, Any]]:
-        result: dict[str, dict[str, Any]] = defaultdict(dict)
-        for status_code, link in links.get_all_links(operation):
-            if isinstance(link, Ok):
-                name = link.ok().name
-            else:
-                name = link.err().name
-            result[status_code][name] = link
-
-        return result
 
     def get_tags(self, operation: APIOperation) -> list[str] | None:
         return operation.definition.raw.get("tags")
@@ -591,14 +566,6 @@ class BaseOpenAPISchema(BaseSchema):
         _maybe_raise_one_or_more(failures)
         return None  # explicitly return None for mypy
 
-    @contextmanager
-    def _validating_response(self, scopes: list[str]) -> Generator[ConvertingResolver, None, None]:
-        resolver = ConvertingResolver(
-            self.location or "", self.raw_schema, nullable_keyword=self.adapter.nullable_keyword
-        )
-        with in_scopes(resolver, scopes):
-            yield resolver
-
 
 def _maybe_raise_one_or_more(failures: list[Failure]) -> None:
     if not failures:
@@ -615,15 +582,6 @@ def in_scope(resolver: jsonschema.RefResolver, scope: str) -> Generator[None, No
         yield
     finally:
         resolver.pop_scope()
-
-
-@contextmanager
-def in_scopes(resolver: jsonschema.RefResolver, scopes: list[str]) -> Generator[None, None, None]:
-    """Push all available scopes into the resolver."""
-    with ExitStack() as stack:
-        for scope in scopes:
-            stack.enter_context(in_scope(resolver, scope))
-        yield
 
 
 @dataclass
@@ -680,10 +638,7 @@ OPENAPI_20_DEFAULT_FORM_MEDIA_TYPE = "multipart/form-data"
 class SwaggerV20(BaseOpenAPISchema):
     example_field = "x-example"
     examples_field = "x-examples"
-    header_required_field = "x-required"
     security = SwaggerSecurityProcessor()
-    component_locations: ClassVar[tuple[tuple[str, ...], ...]] = (("definitions",),)
-    links_field = "x-links"
     _path_parameter_template = {"in": "path", "required": True, "type": "string"}
 
     def __post_init__(self) -> None:
@@ -844,10 +799,7 @@ class SwaggerV20(BaseOpenAPISchema):
 class OpenApi30(SwaggerV20):
     example_field = "example"
     examples_field = "examples"
-    header_required_field = "required"
     security = OpenAPISecurityProcessor()
-    component_locations = (("components", "schemas"),)
-    links_field = "links"
     _path_parameter_template = {"in": "path", "required": True, "schema": {"type": "string"}}
 
     def __post_init__(self) -> None: